usart_command_t()
Command to set tau values. Tested in debugger. Untested in real world.

+++ include/timers.h
@@ -33,14 +33,14 @@
 // ===========================================================
 // Function Prototypes
 // ===========================================================
-void timers_master_init(float period_us);
+void timers_master_init(uint32_t period_us);
 
 void timers_tau_init(	volatile uint16_t *addr_ccN,
 						volatile uint8_t *addr_ctrlb,
 						volatile uint8_t *addr_intctrlb,
 						uint8_t capture_ch_bm,
 						uint8_t interrupt_level_bm,
-						float tau_us
+						uint32_t tau_us
 					);
 
 void timers_init_clock(	volatile uint16_t *addr_per,
